+++ test/parser.test.ts
@@ -11,6 +11,7 @@ import * as assert from 'assert';
 import * as vscode from 'vscode';
 import * as myExtension from '../extension';
 import * as parser from '../src/cmd_line/parser';
+import * as node from '../src/cmd_line/node';
 import * as token from '../src/cmd_line/token';
 
 suite("Cmd line tests - parser", () => {
@@ -22,27 +23,27 @@ suite("Cmd line tests - parser", () => {
 
 	// TODO: Range tests follow -- should prolly create a suite for this
 	test("can parse left - dot", () => {
-		var cmd : parser.CommandLine = parser.parse(".");
+		var cmd : node.CommandLine = parser.parse(".");
 		assert.equal(cmd.range.left[0].type, token.TokenType.Dot);
 	});
 
 	test("can parse left - dollar", () => {
-		var cmd : parser.CommandLine = parser.parse("$");
+		var cmd : node.CommandLine = parser.parse("$");
 		assert.equal(cmd.range.left[0].type, token.TokenType.Dollar);
 	});	
 
 	test("can parse left - percent", () => {
-		var cmd : parser.CommandLine = parser.parse("%");
+		var cmd : node.CommandLine = parser.parse("%");
 		assert.equal(cmd.range.left[0].type, token.TokenType.Percent);
 	});		
 
 	test("can parse separator - comma", () => {
-		var cmd : parser.CommandLine = parser.parse(",");
+		var cmd : node.CommandLine = parser.parse(",");
 		assert.equal(cmd.range.separator.type, token.TokenType.Comma);
 	});
 
 	test("can parse right - dollar", () => {
-		var cmd : parser.CommandLine = parser.parse(",$");
+		var cmd : node.CommandLine = parser.parse(",$");
 		assert.equal(cmd.range.left.length, 0);
 		assert.equal(cmd.range.right.length, 1);
 		assert.equal(cmd.range.right[0].type, token.TokenType.Dollar, "unexpected token");
